store/requirements.txt
Django==3.2.14
django-environ==0.10.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5